+++ ext/java/nokogiri/XmlNode.java
@@ -305,6 +305,14 @@ public class XmlNode extends RubyObject
     IRubyObject name = args[0];
     IRubyObject doc = args[1];
 
+    if (!(doc instanceof XmlNode)) {
+      throw context.runtime.newArgumentError("document must be a Nokogiri::XML::Node");
+    }
+    if (!(doc instanceof XmlDocument)) {
+      // TODO: deprecate allowing Node
+      context.runtime.getWarnings().warn("Passing a Node as the second parameter to Node.new is deprecated. Please pass a Document instead, or prefer an alternative constructor like Node#add_child. This will become an error in a future release of Nokogiri.");
+    }
+
     Document document = asXmlNode(context, doc).getOwnerDocument();
     if (document == null) {
       throw context.runtime.newArgumentError("node must have owner document");